+++ figma2sdui/src/commonMain/kotlin/com/sacada/figma2sdui/data/nodes/properties/Vector.kt
@@ -3,6 +3,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Vector(
-    val x: Int,
-    val y: Int
+    val x: Double,
+    val y: Double
 )
